backend/requirements-Copy1.txt
# 核心依赖
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic[email]==2.5.0
python-multipart==0.0.6

# 数据库
sqlalchemy==2.0.23
alembic==1.13.0
psycopg2-binary==2.9.9
asyncpg==0.29.0

# 缓存
redis==5.0.1
aioredis==2.0.1

# 队列
celery==5.3.4
flower==2.0.1

# 认证安全
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
cryptography==41.0.7

# AI/ML
autogen==0.10.1
openai==1.109.1
tiktoken==0.5.1
langchain==0.0.349

# 代码分析
bandit==1.7.5
safety==2.4.0b1
semgrep==1.44.0
pylint==3.0.3
mypy==1.7.1
black==23.11.0

# 工具库
loguru==0.7.2
aiofiles==23.2.1
python-dotenv==1.0.0
click==8.1.7
pygments==2.17.2
chardet==5.2.0
cchardet==2.1.7

# 数据处理
pandas==2.1.3
numpy==1.26.2
joblib==1.3.2

# WebSocket
websockets==12.0

# 监控
prometheus-client==0.19.0
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-exporter-otlp==1.21.0

# 开发工具
pre-commit==3.5.0